fix(rust): drop the no-op rlimit cast instead of allowing the lint
Sikemux only builds for macOS, where rlim_t is already u64, so the cast
and the clippy allow guarding it can both go.

+++ src-tauri/src/system.rs
@@ -550,17 +550,14 @@ fn current_fd_count() -> Option<usize> {
     None
 }
 
-// `rlim_t` is `u64` on macOS but narrower on some other Unix targets, so the
-// cast is needed there even though it is a no-op here.
 #[cfg(unix)]
-#[allow(clippy::unnecessary_cast)]
 fn current_fd_limit() -> (Option<u64>, Option<u64>) {
     unsafe {
         let mut lim = std::mem::zeroed::<libc::rlimit>();
         if libc::getrlimit(libc::RLIMIT_NOFILE, &mut lim) != 0 {
             return (None, None);
         }
-        (Some(lim.rlim_cur as u64), Some(lim.rlim_max as u64))
+        (Some(lim.rlim_cur), Some(lim.rlim_max))
     }
 }
